# docs/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Points Redemption API",
        "description": "Admin API for managing customers, processing point redemptions, and reconciling transactions.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local"
        }
    ],
    "paths": {
...
        "/api/reports/summary": {
            "get": {
                "tags": [
                    "Reports"
                ],
                "summary": "Get summary report",
                "operationId": "3a2f881744aa1af17114740501546356",
                "parameters": [
                    {
                        "name": "from",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        },
                        "example": "2026-06-01"
                    },
                    {
                        "name": "to",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        },
                        "example": "2026-06-30"
                    },
                    {
                        "name": "customer_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 11
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "pending",
                                "successful",
                                "failed"
                            ]
                        }
                    },
                    {
                        "name": "service_type",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "airtime",
                                "data",
                                "electricity",
                                "cable_tv"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Summary report retrieved",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "total_redemptions": {
                                                    "type": "integer",
                                                    "example": 15
                                                },
                                                "total_points_deducted": {
                                                    "type": "integer",
                                                    "example": 4500
                                                },
                                                "total_amount": {
                                                    "type": "integer",
                                                    "example": 450000
                                                },
                                                "success_rate": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 86.67
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/reports/export": {
            "get": {
                "tags": [
                    "Reports"
                ],
                "summary": "Export report to CSV",
                "operationId": "565ce0189bd2b9916ecb8a80b4b0d1f7",
                "parameters": [
                    {
                        "name": "from",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        },
                        "example": "2026-06-01"
                    },
                    {
                        "name": "to",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        },
                        "example": "2026-06-30"
                    },
                    {
                        "name": "customer_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 11
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "pending",
                                "successful",
                                "failed"
                            ]
                        }
                    },
                    {
                        "name": "service_type",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "airtime",
                                "data",
                                "electricity",
                                "cable_tv"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "CSV file stream",
                        "content": {
                            "text/csv": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}